URL updates + add dropdown
