+++ src/components/contests/Hero.tsx
@@ -2,7 +2,7 @@ import * as React from "react"
 // @ts-ignore
 import HeroImg from "../../images/ContestsHero"
 import Link from "next/link"
-import EmailForm from "../classes/EmailForm"
+import EmailForm from "../video-classes/EmailForm"
 
 export default function Hero() {
   return (
