+++ src/components/video-classes/Hero.tsx
