+++ src/components/video-classes/NotLowQuality.tsx
