+++ src/components/video-classes/registration/Divider.tsx
