+++ src/components/video-classes/registration/ErrorCard.tsx
