+++ src/components/video-classes/registration/FinancialAidSubmissionSection.tsx
@@ -103,7 +103,7 @@ export default function FinancialAidSubmissionSection({
                   const ref = firebase
                     .firestore()
                     .collection("classes-registration")
-                    .doc("2021october")
+                    .doc("usacobronze")
                     .collection("registrations")
                     .doc()
                   setRegistrationId(ref.id)
