+++ src/components/video-classes/registration/SubmittingCard.tsx
