+++ src/pages/classes/admin/view-registrations.tsx
@@ -108,6 +108,7 @@ export default function ViewRegistrationPage() {
             "OjLKRTTzNyQgMifAExQKUA4MtfF2",
             "v8NK8mHCZnbPQKaPnEs5lKNc3rv2",
             "BKFOe33Ym7Pc7aQuET57MiljpF03",
+            "YF9ObmH1SUR1MKJGTrO8DfBQUG13"
           ].includes(user.uid)
         ) {
           setHasPermission(true)
@@ -115,7 +116,7 @@ export default function ViewRegistrationPage() {
             firebase
               .firestore()
               .collection("classes-registration")
-              .doc("usacobronze")
+              .doc("2021october")
               .collection("registrations")
               .onSnapshot(snapshot => {
                 const newRegistrations = []
@@ -176,7 +177,9 @@ export default function ViewRegistrationPage() {
           <h1 className={"text-4xl font-bold tracking-tight leading-9"}>
             Error 404: Page Not Found
           </h1>
-          <Link href={"/"}>
+          <Link
+            href={"/"}
+          >
             <a className={"text-2xl text-blue-600 hover:underline pt-4 block"}>
               Go Home
             </a>
@@ -303,9 +306,6 @@ export default function ViewRegistrationPage() {
                               ? "Java"
                               : reg.data.personalInfo.preferredLanguage == "cpp"
                               ? "C++"
-                              : reg.data.personalInfo.preferredLanguage ==
-                                "python"
-                              ? "Python"
                               : "ERROR"}
                           </p>
                         </div>
@@ -473,10 +473,7 @@ export default function ViewRegistrationPage() {
                           {detailModalRegistrationData.personalInfo
                             .preferredLanguage === "java"
                             ? "Java"
-                            : detailModalRegistrationData.personalInfo
-                                .preferredLanguage === "cpp"
-                            ? "C++"
-                            : "Python"}
+                            : "C++"}
                         </p>
                         <p className=" text-gray-900">
                           <b>Referrer:</b>{" "}
@@ -523,7 +520,7 @@ export default function ViewRegistrationPage() {
                           if (
                             !confirm(
                               detailModalRegistrationData
-                                .financialAidApplication.amount !== 25
+                                .financialAidApplication.amount !== 100
                                 ? "Are you sure you want to grant full financial aid? The applicant specified that they would be willing to pay part of the fee.  This action is irreversible."
                                 : "Are you sure you would like to grant full financial aid? This action is irreversible."
                             )
@@ -538,7 +535,7 @@ export default function ViewRegistrationPage() {
                           setDetailModalFASubmittingApproval(true)
                           firebase
                             .functions()
-                            .httpsCallable("cpiclasses-approveFinancialAid")({
+                            .httpsCallable("cpiclasses-approveLiveFinancialAid")({
                               registrationId: detailModalRegistrationId,
                               email:
                                 detailModalRegistrationData.personalInfo.email,
@@ -591,7 +588,7 @@ export default function ViewRegistrationPage() {
                           firebase
                             .firestore()
                             .collection("classes-registration")
-                            .doc("usacobronze")
+                            .doc("2021october")
                             .collection("registrations")
                             .doc(detailModalRegistrationId)
                             .update({
